+++ Editor/Editors/TextureEditor/TextureEditor.cs
@@ -1,19 +1,29 @@
 ﻿using Editor.Common;
 using Editor.Common.Enums;
 using Editor.Content;
+using Editor.Utilities;
 using System.Diagnostics;
 using System.IO;
+using System.Windows;
+using System.Windows.Media.Imaging;
 
 namespace Editor.Editors
 {
     public class TextureEditor : ViewModelBase, IAssetEditor
     {
+        private readonly List<List<List<BitmapSource>>> _sliceBitmaps = new();
+
         private AssetEditorState _state;
         private Texture _texture;
+        private List<List<List<Slice>>> _slices;
+        private Point _panOffset;
+        private double _scaleFactor = 1.0;
 
         public Guid AssetGuid { get; private set; }
 
         Asset IAssetEditor.Asset => Texture;
+        public BitmapSource SelectedSliceBitmap => _sliceBitmaps.ElementAtOrDefault(0)?.ElementAtOrDefault(0)?.ElementAtOrDefault(0);
+        public Slice SelectedSlice => Texture?.Slices?.ElementAtOrDefault(0)?.ElementAtOrDefault(0)?.ElementAtOrDefault(0);
 
         public AssetEditorState State
         {
@@ -36,6 +46,33 @@ namespace Editor.Editors
                 {
                     _texture = value;
                     OnPropertyChanged(nameof(Texture));
+                    SetSelectedBitmap();
+                }
+            }
+        }
+
+        public Point PanOffset
+        {
+            get => _panOffset;
+            set
+            {
+                if (value != _panOffset)
+                {
+                    _panOffset = value;
+                    OnPropertyChanged(nameof(PanOffset));
+                }
+            }
+        }
+
+        public double ScaleFactor
+        {
+            get => _scaleFactor;
+            set
+            {
+                if (value != _scaleFactor)
+                {
+                    _scaleFactor = value;
+                    OnPropertyChanged(nameof(ScaleFactor));
                 }
             }
         }
@@ -59,11 +96,13 @@ namespace Editor.Editors
                     texture.Load(info.FullPath);
                 });
 
+                await SetMipmapsAsync(texture);
+
                 Texture = texture;
             }
             catch(Exception ex)
             {
-                Debug.Write(ex.Message);
+                Debug.WriteLine(ex.Message);
                 Debug.WriteLine($"Failed to set texture for use in texture editor. File: {info.FullPath}");
                 Texture = new();
             }
@@ -71,6 +110,59 @@ namespace Editor.Editors
             {
                 State = AssetEditorState.DONE;
             }
+        }
+
+        private async Task SetMipmapsAsync(Texture texture)
+        {
+            try
+            {
+                await Task.Run(() =>
+                {
+                    _slices = texture.Slices;
+                });
+
+                Debug.Assert(_slices?.Any() == true && _slices.First()?.Any() == true);
+
+                GenerateSliceBitMaps(texture.IsNormalMap);
+                OnPropertyChanged(nameof(Texture));
+            }
+            catch(Exception ex)
+            {
+                Debug.WriteLine(ex.Message);
+                Debug.WriteLine($"Failed to load mipmaps from {texture.FileName}");
+            }
+        }
+
+        private void GenerateSliceBitMaps(bool isNormalMap)
+        {
+            _sliceBitmaps.Clear();
+
+            foreach(var arraySlice in _slices)
+            {
+                List<List<BitmapSource>> mipmapsBitmaps = new();
+
+                foreach(var mipLevel in arraySlice)
+                {
+                    List<BitmapSource> sliceBitmap = new();
+
+                    foreach(var slice in mipLevel)
+                    {
+                        var image = BitmapHelper.ImageFromSlice(slice, isNormalMap);
+
+                        Debug.Assert(image is not null);
+
+                        sliceBitmap.Add(image);
+                    }
+                    mipmapsBitmaps.Add(sliceBitmap);
+                }
+                _sliceBitmaps.Add(mipmapsBitmaps);
+            }
+        }
+
+        private void SetSelectedBitmap()
+        {
+            OnPropertyChanged(nameof(SelectedSliceBitmap));
+            OnPropertyChanged(nameof(SelectedSlice));
         }
     }
 }