+++ Editor/Utilities/MathUtilities.cs
@@ -7,12 +7,20 @@ namespace Editor.Utilities
         public static float Epsilon => .00001f;
 
         public static bool IsEqual(this float value, float other) => Math.Abs(value - other) < Epsilon;
+        public static bool IsEqual(this double value, double other) => Math.Abs(value - other) < Epsilon;
 
         public static bool IsEqual(this float? value, float? other)
         {
             if (!value.HasValue || !other.HasValue) return false;
 
             return IsEqual(value.Value, other.Value);
+        }
+
+        public static bool IsEqual(this double? value, double? other)
+        {
+            if (!value.HasValue || !other.HasValue) return false;
+
+            return IsEqual(value, other);
         }
 
         public static long AlignSizeUp(long size, long alignment)